examples/cloud/essence/requirements.txt
# Core bitHuman SDK
bithuman>=0.5.25

# Environment management
python-dotenv~=1.1

# LiveKit agents with all required plugins
livekit-agents[openai,bithuman,silero]~=1.2.11

# Development tools (optional)
ruff>=0.1.0
